+++ skyrim-platform/src/platform_se/skyrim_platform/PapyrusTESEvents.h
@@ -1,12 +1,8 @@
 #pragma once
 #include <RE/ScriptEventSourceHolder.h>
+#include <RE/TESObjectREFR.h>
 
-namespace PapyrusSendEvent {
-template <class T>
-void Emit(T* event);
-void TestEvent();
-void Hit(RE::TESHitEvent* event);
-void SpellCast(RE::TESSpellCastEvent* event);
+namespace PapyrusTESEvents {
 
 struct TESSpellCastEvent
 {
@@ -18,5 +14,4 @@ public:
   RE::FormID spell;
 };
 static_assert(sizeof(TESSpellCastEvent) == 0x10);
-
 }
